iox-#13 add config file parser and some refactorings

+++ iceoryx_examples/iceperf/topic_data.hpp
@@ -18,6 +18,6 @@
 
 struct PerfTopic
 {
-    uint64_t payloadSize{0};
+    uint32_t payloadSize{0};
     bool run{true};
 };
